Update COWS
Try to implement cows, but i dont know if this is a good idea. i can go to rest service, but can't use DB, because my db don't have a COWS? Why should this help to setup a COWS on rest if it need to be set on the DB instead?

Link StackOverFlow = https://stackoverflow.com/questions/44379560/how-to-enable-cors-in-asp-net-core-webapi?page=1&tab=votes#tab-top